Remove rxjava from app
