+++ main/src/main/kotlin/io/goooler/demoapp/main/repository/MainCommonRepository.kt
@@ -6,7 +6,6 @@ import io.goooler.demoapp.common.type.CommonConstants
 import io.goooler.demoapp.main.api.MainCommonApi
 import io.goooler.demoapp.main.bean.MainRepoListBean
 import io.goooler.demoapp.main.db.MainCommonDao
-import io.reactivex.rxjava3.core.Single
 import javax.inject.Inject
 
 class MainCommonRepository @Inject constructor(
@@ -25,12 +24,6 @@ class MainCommonRepository @Inject constructor(
     )
     return api.getRepoListWithCr(user, params)
   }
-
-  fun getRepoListWithRx(
-    user: String,
-    @IntRange(from = 1) page: Int = 1,
-    @IntRange(from = 1) pageSize: Int = CommonConstants.DEFAULT_PAGE_SIZE
-  ): Single<List<MainRepoListBean>> = api.getRepoListWithRx(user, page, pageSize)
 
   suspend fun getRepoListFromDb(ownerName: String): List<MainRepoListBean> =
     dao.getRepoList(ownerName)
