+++ main/src/main/kotlin/io/goooler/demoapp/main/vm/MainHomeViewModel.kt
@@ -3,7 +3,7 @@ package io.goooler.demoapp.main.vm
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.goooler.demoapp.base.util.defaultAsync
-import io.goooler.demoapp.common.base.theme.BaseRxViewModel
+import io.goooler.demoapp.common.base.theme.BaseThemeViewModel
 import io.goooler.demoapp.common.util.showToast
 import io.goooler.demoapp.main.bean.MainRepoListBean
 import io.goooler.demoapp.main.repository.MainCommonRepository
@@ -25,7 +25,7 @@ import kotlinx.coroutines.launch
 
 @HiltViewModel
 class MainHomeViewModel @Inject constructor(private val repository: MainCommonRepository) :
-  BaseRxViewModel() {
+  BaseThemeViewModel() {
 
   private val _title = MutableStateFlow("")
   val title: StateFlow<String> = _title
